Add Koin navigation3 wiring
