+++ app/build.gradle
@@ -98,6 +98,9 @@ dependencies {
     implementation libs.navigation3.runtime
     implementation libs.navigation3.ui
     implementation libs.lifecycle.viewmodel.navigation3
+    implementation libs.koin.android
+    implementation libs.koin.androidx.compose
+    implementation libs.koin.compose.navigation3
 
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     testImplementation libs.kotlin.stdlib
